// src/api/generated/model/productResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Ecommerce Copa API
 * API backend do Ecommerce Copa com autenticação por sessão, catálogo público e CRUD administrativo.
 * OpenAPI spec version: 1.0.0
 */
import type { CategoryResponse } from "./categoryResponse.ts";
import type { ProductResponseSchema } from "./productResponseSchema.ts";

export interface ProductResponse {
	id?: string;
	category?: CategoryResponse;
	/** Schema de seletores do produto. */
	schema?: ProductResponseSchema;
	skuCount?: number;
}