// src/api/generated/model/skuResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Ecommerce Copa API
 * API backend do Ecommerce Copa com autenticação por sessão, catálogo público e CRUD administrativo.
 * OpenAPI spec version: 1.0.0
 */
import type { SkuResponseAttributes } from "./skuResponseAttributes.ts";
import type { TagResponse } from "./tagResponse.ts";

export interface SkuResponse {
	id?: string;
	productId?: string;
	title?: string;
	description?: string;
	price?: number;
	originalPrice?: number;
	photo?: string;
	stock?: number;
	/** Valores dos atributos do SKU. */
	attributes?: SkuResponseAttributes;
	tags?: TagResponse[];
}